requirements.txt
flask
openai-whisper
pyTelegramBotAPI
pymysql
